# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "FlatWhiteUni Server REST API",
    "description": "FlatWhiteUni Project Server",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8081",
  "basePath": "/api",
  "tags": [
    {
      "name": "FlatWhiteUni",
      "description": "API for FlatWhiteUni server in the system"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/universities": {
      "post": {
        "tags": [
          "Universities"
        ],
        "summary": "Add a new university in system",
        "description": "Create new university in system",
        "parameters": [
          {
            "name": "university",
            "in": "body",
            "description": "university that we want to create",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New university is created",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Universities"
        ],
        "summary": "Get all universities in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Universities"
            }
          }
        }
      }
    },
    "/universities/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of university that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Universities"
        ],
        "summary": "Get university with given ID",
        "responses": {
          "200": {
            "description": "University is found",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete university with given ID",
        "tags": [
          "Universities"
        ],
        "responses": {
          "200": {
            "description": "University is deleted",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        }
      },
      "put": {
        "summary": "Update university with give ID",
        "tags": [
          "Universities"
        ],
        "parameters": [
          {
            "name": "university",
            "in": "body",
            "description": "university with new values of properties",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "university is updated",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        }
      }
    },
    "/universities/name/{name}": {
      "parameters": [
        {
          "name": "name",
          "in": "path",
          "required": true,
          "description": "Name of university that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Universities"
        ],
        "summary": "Get university with given name",
        "responses": {
          "200": {
            "description": "University is found",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        }
      }
    },
    "/universities/city/{city}": {
      "parameters": [
        {
          "name": "city",
          "in": "path",
          "required": true,
          "description": "City of universities that we want to find",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "Universities"
        ],
        "summary": "Get universities with given city",
        "responses": {
          "200": {
            "description": "University is found",
            "schema": {
              "$ref": "#/definitions/University"
            }
          }
        }
      }
    },

    "/schools":{
      "post": {
        "tags": [
          "Schools"
        ],
        "summary": "Add a new school in system",
        "description": "Create new school in system",
        "parameters": [
          {
            "name": "school",
            "in": "body",
            "description": "school that we want to create",
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "New school is created",
            "schema": {
              "$ref": "#/definitions/School"
            }
          }
        }
      },
      "get": {
        "tags": [
          "Schools"
        ],
        "summary": "Get all schools in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Schools"
            }
          }
        }
      }
    }
  },

  "definitions": {
    
    "University": {
      "required":[
        "name",
        "city",
        "_id"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "city": {
          "type": "string"
        }
    }
  },

  "Universities": {
      "type": "array",
      "$ref": "#/definitions/University"
    },

  "School": {
      "required":[
        "name",
        "universityName",
        "_id"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "name": {
          "type": "string",
          "uniqueItems": true
        },
        "universityName": {
          "type": "string"
        }
    }
  },

  "Schools": {
      "type": "array",
      "$ref": "#/definitions/School"
    }

}
}